archival/requirements.txt
azure-cosmos
azure-storage-blob
